Safari-609.1.11.1/LayoutTests/media/track/in-band/track-in-band-srt-mkv-addtrack-expected.txt


Testing text track "addtrack" events.

EVENT(addtrack)
EXPECTED (video.textTracks contains 'event.track') OK

EVENT(addtrack)
EXPECTED (video.textTracks contains 'event.track') OK

EVENT(addtrack)
EXPECTED (video.textTracks contains 'event.track') OK

EVENT(addtrack)
EXPECTED (video.textTracks contains 'event.track') OK

EVENT(canplaythrough)
END OF TEST
